+++ web-ui/src/components/RootLayout.tsx
@@ -73,7 +73,7 @@ export function RootLayout() {
             © 2025 Adamax-Works
           </Typography>
           <Typography variant="caption" display="block" color="text.secondary">
-            v2.5.2
+            v2.5.3
           </Typography>
         </Box>
       </Box>